resolving conflicts
